Compile web app as phonegap app
I've had to generate the __layout.hbs file using a jade file so that I can also generate the index.html used for the phonegap app.

The index.html init js uses the two new settings I added to rendr: clientUrlBase and renderOnFirstRoute. The clientUrlBase is generated in Gruntfile.js by looking up your IPv4 address (currently for adaptor en0) and getting the port using the same derivation as index.js. This is done so that as long as your phone/emulator and laptop are running on the same network the phonegap app should use the rendr server running on your computer for its api calls.

I have setup an email address (rendrexampleapp@gmail.com) and phonegap build account (both use the same password). The phonegap build account is then used to compile the web app and an android.apk is downloaded.

It currently only downloads android.apk because: to get an iOS build I would need to attach my iOS dev keys to the phonegap build account which I am not happy with; and I am not interested in any other platforms but those two.